+++ CHANGELOG.md
@@ -4,10 +4,17 @@
 
 ### Added
 - [CodeEngine] Compatible runtimes between knative and CE
+- [CodeEngine] runtime name regex verification
+- [Standalone] Append installation logs into /tmp/lithops/proxy.log
 
 ### Changed
 - [Localhost] Run functions in processes instead of threads in Windows
 - [CodeEngine] Reduced payload size
+- [CodeEngine] Changed way to obtain modules preinstalls
+- [Core] Updated logging
+
+### Fixed
+- [Core] Fixed Cloudpickle 1.6 modules detection
 
 
 ## [v2.2.11]